requirements.txt
uvicorn[standard]==0.23.2
django==4.2
djangorestframework==3.14.0
django-filter==23.4
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
tzlocal==5.2
sentry-sdk==1.38.0
